Restructured Auto_Snap.py for better use of numpy, multiprocessing, and image scale caching.
Added 'Tab to skip GUI' option to the GUI, allowing for the current transform to be applied to the rest of the images.

Modified DepMan to only install packages. Needs additional work for compatibility.

Added modularity multiprocessing support to Post_Processing.

Added instructions to Stitching_main.

Wrote ipython demo of the automatic alignment. Not sure if that will render on GitHub.